IKHSS7-18590 - Change the default USB Device Class
Default USB Device Class, SubClass and Protocol is getting reported
as Vendor Specific. In certain cases the driver installation can fail
because there is no Vendor specified Composite Driver for that mode.
Change the Class, SubClass And Protocol to be specified at the interface
level so that the standard Composite driver is always loaded.

Change-Id: I8bc2a91037843c90d6c9ee11b3fb0630c04c3c56
Reviewed-on: http://gerrit.pcs.mot.com/384015
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Neil Patel <NPATEL@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>