ASoC: wcd9310: Add support for MIC BIAS other than 1.8 volts
When MBHC is set to use microphone bias other than 1.8 volts, codec driver
needs to calculate two sets of thresholds. This is required because the
mic bias can be powered from two different sources. If vddio is used as
supply, then the maximum microphone bias voltage can only be 1.8 volts.
If MBHC originally was using other microphone voltage, the thresholds will
have to be changed when mic bias power is switched to vddio supply.

Change-Id: Ic5b27416f0842249c4c3b68b8442c5077f5194b2
Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>
Reviewed-on: http://gerrit.pcs.mot.com/391490
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>