IKHSS7-21020 HDCP Compliance Test Fixes
This change fixes a number of race conditions in the HDMI/HDCP
driver.  Many of the race conditions were in the base code, but
potentially not uncovered until the "Support for non-HDCP"
devices change went in.  The race conditions are mostly related
to the re-auth timer not being stopped when a cable disconnection
occurs.  Others relate to phantom interrupts being caused by the
HPD 5V supply being turned off and on again when the HDMI output
is disabled.

Change-Id: I72b081cdc39d82ab3224eae07d9fbede0ab1dfc2
Signed-off-by: Steven Fischer <Steven.Fischer@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/386898
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Sumit Bhojesing Jamadar <sumit@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>