IKHSS7-20520:VBUS IRQs fire several times with USB cable attached
This patch includes the fix given by qualcom for IKHSS7-20520 and
removes the temporary fix for the same.

From: David Keitel <dkeitel@codeaurora.org>
When the usb_suspend bit is set and an the CHG_GONE
interrupt is triggered the ovp fets are opened to
enforce a usb unplug detection.

Fix this by checking the CHG_USB_SUSPEND bit and do not
open the USB OVP FET.

Change-Id: I08eb93818d1ece95e05fb3f46a7f6da4994e0e54
Signed-off-by: Swetha <a21837@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/387163
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Neil Patel <NPATEL@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Kenneth Kessler <w65789@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>