IKHSS7-16591: Phone shows 99% at 4.35V
This error was created because the tenths of percent place is not used in the
State of Charge.  Integer maths means that 99.9% is 99%.  This change fixes
that by rounding all remainders up to the whole percent so 99.1% to 100% is
100%.  Also we had to modify the real FCC correction so that there is not an
inflated FCC by only allowing corrections if the new is smaller then the
present.

Change-Id: I39db0027616bd73875594afc21958e096e219f81
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/388344
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Vidyadhar Gurram <Vidyadhar.Gurram@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Robert Johnson <WBJ002@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>