IKHSS7-15111 Loudspeaker pop
Qualcomm Patch:
You can find it in the Hardware document -80-VP942-89 : End to End Enable Sequence section,
you can find this information.

0x1CA; 0x1F; 0x85

In the codec driver, Currently we are applying 16 msec of delay while enalbing the each LINEOUT PA.
For Mono Differential Settings this delay will be 16*2 = 32 msec, but actually we need a delay of
64 msec. Please change this dealy from 16 msec to 32 msec if you are going to use the Mono Differenti
But for the Stereo Differnetial 16 msec is sufficent as we have to enable four Lineout PA's - 16*4 =64.

Change-Id: Ife3a41167c67522f1078182362692edf795f18c4
Signed-off-by: Harini Peddibhotla <hpeddib1@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/388832
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Shivank Sinha <shivank.sinha@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>