IKHSS7-17576: error handling for cards stuck in programming state
Attempt to deal with cards that get stuck in the programming state
during a write by sending them a stop command and then retrying the
transfer.  This condition is very rare and so this approach may need
additional tweaking.

Clean up the card resume retries by scaling-up the "settling" delays
with each retry.  Also add a log when the retry worked to ease the
minds of the test team.

Change-Id: Ia79b6110a19e52aad1618341087f2e11216a9894
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/388879
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>