IKHSS7-12343: Unbalanced Enable of BTM IRQs
configure_btm was being called twice which enabled the btm irqs.
All that needs to be done after eprom read is change the warm and cool
temperature.  This change sets the global for warm  nd cool thresholds and then
schedules the btm_configure work to handle the change without enabling the irqs

Change-Id: I8772209a11729be98e558e70bda34376f4f79aee
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/388891
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>