IKHSS7-22185 msm/gpiomux: set gpiomux by devtree
The MSM gpiomux should be set before msmgpio module init. The kernel already
set some default settings, but we need to set a different value for different
product.

This patch will add an interface to read the gpiomux setting from devtree to
override the default setting in kernel.

The devtree setting format for gpiomux is as follows:
       <node name="System@0">
    +
    +    <node name="IOMUX@0" type="Qualcomm,MSM8960,IOMUX">
    +      <property name="settings" value="{
    +        {GPIO002, ACTIVE,     MODE1, PULL_DISABLE, 2MA, GPIOMUX_IN},
    +        {GPIO002, SUSPENDED,  MODE1, PULL_DISABLE, 2MA, GPIOMUX_IN}
    +      }" />
    +    </node>

Change-Id: I93145676f8422fe17d7b1550d56919b955c9e0e5
Signed-off-by: Lianwei Wang <lian-wei.wang@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/389264
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Guo-Jian Chen <A21757@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>