IKQCOMSD1-1436: Audio Route Support To EMU Headset
This chnage is to support EMU audio for factory testing. In the
factory, when the AUDIO test command is sent, the phone SW has to
route audio played by media file to the EMU headset

For this the factory need to set the test mode to MODE_F_AUDIO and play
Audio. The reason for teh new mode is to make sure that the detection
algorithm is not invoked while testing Audio. The D+ & D- lines are
routed to Audio mode, and whenever the audio is played, it is routed over
D+ & D- lines.

Change-Id: I3d7accdda13fd13adea289251297daf67cc401e1
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/389712
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>