IKHSS7-19289: Add support for unpowered 200 kohm charger case.
Legacy Motorola chargers use a 200 kohm resistor to ground
on the ID line. If this charger is inserted when unpowered,
it will be detected as an unauthenticated desk dock with
headset attached. The phone will keep power applied to the
charger the entire time it is connected even after the charger
becomes powered. The only way to recover is to re-insert the
powered charger.

To resolve this issue, the unpowered charger will be detected
as nothing connected. Then, if/when power is applied to the
connected charger, the detection code will properly detect
the charger. Note that this change is also being applied
for the case of a 100 kohm charger to cover all corner cases.

Change-Id: I82780a1f14426f7cfb4696bdf489b5280c0f3649
Signed-off-by: Jason Knopsnyder <w19069@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/390810
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Vidyadhar Gurram <Vidyadhar.Gurram@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>