IKHSS7-23607: Reduce the delay in the read bit function for 1-wire
For the EEPROM read failures we reduced the delay when reading a bit
from the slave, i.e. delay after driving the bus low & delay after
releasing the bus form 6 to 4 usec and 9 to 6 usec respectively.

The test results were more promising when we further reduced the
delay i.e delay after drive the bus low to 3 usec and delay after
releasing the bus to 4 usec.

Change-Id: I57cae637a54f79d6e0031f7af8c6adeb278767eb
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/392420
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>