IKHSS7-23671: android: binder: reduce memory for binder debugfs
Reduce the amount of memory the binder debugfs seq_file's need
to output the large amount of data they provide.  This breaks up
the individual reads into smaller objects to limit the kmalloc'd
memory to a PAGE.  This helps reduce OOM conditions when these
files are read in mildly fragmented systems.

Change-Id: I368573914f73975321270ea1cafe3fb3ef7de28a
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/391877
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>