IKHSS7-16869 Fix for lost HDMI HPD interrupts
This change resolves an issue where HPD interrupts were being
lost during an HDMI HW core reset which happens when the output
is being turned on.  The change does two things, first we trap
lost interrupts when the core is reset, and second we reduce
the chance of lost interrupts by not reseting the core when
performing a re-authentication.

Change-Id: Id69ce9ac6a966252dd4fb5a49a76eabb2ae1d9a6
Signed-off-by: Steven Fischer <Steven.Fischer@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/392297
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Sumit Bhojesing Jamadar <sumit@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>