IKHSS7-23775: pm8xxx-adc: Make ADC complete interrupt wakeable
There exists situations when an adc conversion is requested very
early in the resume sequence. These requests fail to complete
because of missing interrupts and eventually cause lockups.

When an interrupt is enabled when the system is in resume
the kernel rejects enabling these interrupts. This keeps the interrupt
disabled and causes missed interrupt. Get around this problem
by not needing to enable the interrupt. IOW, do not call disable_irq
and enable_irq, only request the irq which leaves it in enabled state.

There is a bug in the pmic where if an interrupt is masked/unmasked
it clears the pending status. This causes a nonwakeup interrupt to
be missed as it is unmasked while resuming. Mostly this is okay but
if a adc request is made early in the suspend, it causes it to miss
interrupts. Make the adc interrupt a wakeup interrupt.

The driver initializes its completion only at probe time. This may cause
incorrect end of conversions with out of sync (spurious) interrupts.
Update the code to init a completion just before issuing a conversion.

Another issue was observed where the interrupt's unmask_irq was delayed,
causing another cpu to initiate the next conversion as soon as complete was
called. This lead to a situation where the interrupt for the next request
was triggered and the unmask happened later. Call complete within a
workqueue which guarantees that unmask would have completed and
it is safe to start the next conversion.

Change-Id: I79582060e62695f0070c5f54120e4a0819ae3342
Signed-off-by: Jason Knopsnyder <w19069@motorola.com>
Change-Id: I8f49dbbeee124495a14a480f7e78cb6b5d7102c6
Reviewed-on: http://gerrit.pcs.mot.com/392366
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Kenneth Kessler <w65789@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>