IKHSS7-22324: Stop Adjusting FCC at Full
QCOM design tries to calculate the Full Charge Capacity at 100%. This does not
function properly with the Float Charging Feature Motorola added.  The problem
led to either the FCC Increaseing or Decreasing uncontrollably.  This meant
the Battery Meter could make strange jumps in either direction and the Energy
Full parameter to be wrong.

Change-Id: Id18f1685e9e6f393797d6ffba4fb7dda8ce2528c
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/392406
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>