IKHSS7-24918: Charge to 4.0V in Extended Temperature
The New Qualification Spec allows charging a Battery to 4.0V when in the
extended range(45C to 60C or 0C to -20C).  All Embedded Batteries have been
Qualified so change them to meet this change and increase User Experience.
Default Table was left at 3.8V to keep things as safe as possible.

Change-Id: I4441a67abefe825c6e863168c60b96c2d4b45cbe
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/397989
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>