IKHSS7-19189: enable sd card encryption on qualcomm line
We deploy eCryptFS to handle the sd card encryption,so first,
we need to make eCryptfs module get build into kernel package.

Second,we need to add the corresponding part that processing
the "no_new_encrypted" flag in eCryptFS of kernel,which as a
part of parameters when we mount eCryptFS actually controls
whether new incoming files should be encrypted or not.

Change-Id: I28e14d5e90dbf405d7822d4f84fc30a8fc357f46
Reviewed-on: http://gerrit.pcs.mot.com/396674
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: John Zafiris <John.Zafiris@motorola.com>
Reviewed-by: Dmitri Latypov <Dmitri.Latypov@motorola.com>