IKHSS7-25419: mmc: sdcc: disable runtime suspend
There is a suspected hardware issue with respect to power management
and SD cards.  The 5 second idle timer interacts poorly with the 5-10
log flush rate from aplogd.  The result is that the card goes in and
out of low power every 5 seconds. This change disables runtime power
management to see if this theory holds any water.

Change-Id: I274c09fb1ff022b938e14535f608fe804f257672
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/396841
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>