IKHSS7-24535 Trigger Watchdog when generic_exec_single() softlockup happens
The challedge to resolve the "generic_execi_single()" softlockup is
it's hard to know what the counterpart CPU is doing when one CPU
is locked up. Since MSM8960's Watchdog can dump CPU registgers, if
softlockup trigger watchdog, it may provide more clues.

The watchdog will be triggered only when on CPU waiting for IPI reponse
too long and finally softlockup happens.

This is just for debugging and will be reverted after related issues
are resolved.

Change-Id: I644a8eb7e6850946396b934c51fc1e2d9b71ffca
Signed-off-by: Yi-wei Zhao <gbjc64@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/396617
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Zhi-Ming Yuan <zhi-ming.yuan@motorola.com>