IKHSS7-25801: Add New QCOM BMS Patches
pm8921-charger: fix 10mv error on vdd_max

It was observed that the charging voltage is off by 10mV. The root
cause was the 10mV bit in the vdd_max register adds 10mV to the voltage
instead of reducing the voltage by 10mV.

The current driver thinks it subtracts 10mV. Fix this.

Change-Id: I37949113cb4c25642c873ce0cf39ac9fdeb276e1
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

power: pm8921-bms: add protection against divide by
 zero

There is a unprotected divide by zero in the driver- when full charge
capacity of the battery and unusable charge of the battery are calculated
equal.

In such a case set state of charge to 0% without the actual division.

Change-Id: I57f3333e780396de29435d29c30e54940c657d4e
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

power: pm8921-bms: calibrate xoadc more often

Improve the values of the battery voltage read from xoadc by periodically
calibrating the xoadc.

The xoadc needs to be calibrated every time a charger is plugged in
or plugged out. This is because the reference voltage shifts a bit
when a charger is plugged in.

Provide an api to calibrate xoadc. This api will be used by the charger
driver.

CRs-Fixed: 334825
Change-Id: I09735f6e5c5d4ab8be379fe001fff81c25c95397
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

power: pm8921-bms: add api to simultaneously read
 vbatt and ibatt

The compensation algorithms need simultaneous readings of the battery
voltage (vbatt) and the current flowing through it (ibatt).

One needs to put the BMS in override mode to get simultaneous vbatt
and ibatt readings.

Note that since one would need to sleep for 20 milliseconds after the
BMS is put in override mode the spin_lock is changed to mutex in the bms
driver.

Change-Id: I1bfdc1f4e84d5d3c2cd6502dd66afdfba650cc0d
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

power: pm8921-bms: calculate rbatt only at resume

The battery resistance (rbatt) measurement can only happen when the
current consumption of the device goes very low. This implies that
rbatt should be checked for only while resuming.

Update the driver to check and measure for rbatt only during resume.
As a result separate the functions that read state of charge (soc)
parameters from rbatt parameters. Also note the battery temperature
right before suspend. This will be used for adjusting the rbatt based
on temperature.

Change-Id: I254833f437605d6d64dcd6159edb8440d08383db
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

power: pm8921-bms: calibrate ocv only when it changes

The current driver adjusts the ocv (open circuit voltage) every time
it is asked to do a state of charge calculation. This results in small
errors in state of charge as the calibration values changes.

Update the driver such that it converts a raw ocv value to voltage only
when it changes. For other times it uses the previously calculated ocv
voltage (stored in last_ocv_uv). This implies that calibration data used
for raw to voltage conversion is right around when ocv was taken.

Change-Id: I27832682f77391c4fce65127e25ec456ee1a0fd1
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

power: pm8921-bms: adjust the pon ocv

The power on open circuit voltage (pon ocv) is always reported higher.
Reduce 15mV from the pon ocv.

Also on chips trimmed with revision lower than 9 - there is an additional
error on all the battery voltage readings. Use the trim program version
and the trim register to nullify this error.

CRs-Fixed: 347207
Change-Id: I873045b9832ec33d044f7cd4aae71254d7be774f
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>

Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Change-Id: I17799e3cccf0eb49a731c2b8168be2ae65e03a2a
Reviewed-on: http://gerrit.pcs.mot.com/397741
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>