msm: pm_8x60: Allow tunable residency values for Low power modes
Enable control of Low power mode residency through sysfs. This prevents
a low power mode from being chosen if the time between current and the
next scheduled wakeup is less than the residency value specified.

The residency value (in uSec) for core0 WFI can be set at
/sys/module/pm_8x60/modes/cpu0/wfi/residency
and so on for the combination of cpus and low power modes.

The default value is 0.

Change-Id: I4639f16964450c01663b7216531b6fad3739321f
Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>
Reviewed-on: http://gerrit.pcs.mot.com/398114
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Ravi Chebolu <Ravi.Chebolu@motorola.com>
Reviewed-by: Lian-Wei Wang <a22439@motorola.com>
Reviewed-by: David Ding <dding@motorola.com>