cpufreq: interactive: based hispeed bump on target freq, not actual
For systems that set a common speed for all CPUs, checking current
speed here could bypass the intermediate hispeed bump decision for
this CPU when another CPU was already at hispeed.  This could
result in an overly high setting (for all CPUs) in situations
where all CPUs were about to drop to load levels that map to
hispeed or below.

Change-Id: I186f23dcfc5e2b6336cab8b0327f0c8a9a4482bc
Signed-off-by: Todd Poynor <toddpoynor@google.com>
Reviewed-on: http://gerrit.pcs.mot.com/428394
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Lian-Wei Wang <a22439@motorola.com>
Reviewed-by: Guo-Jian Chen <A21757@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>
Reviewed-by: Srinivas Sripathi <sripathi@motorola.com>