IKMSM8960BP-884: merge Qualcomm patch, SR 00827472
When SMSM interrupt is received, a snapshot of SMSM callback is stored
in a kfifo. The snapshot is stored as
a) SMSM state of each subsystem
b) Wakelock usage information.

Then a work item is scheduled in global workqueue to process the snapshot and
notify the smsm clients. The smsm snapshot is processed as
1) Lock the snapshot
2) Retrieve SMSM states from kfifo and process them
3) Unlock the snapshot
4) Retrieve Wakelock usage information from kfifo

When back-to-back smsm interrupts are received, multiple work items are
scheduled in multiple cores. The first work item performs steps 1, 2 & 3
and got pre-empted. In the meanwhile the second work item starts with
Wakelock usage information and reports wrong smsm state to the SMSM clients.

Solution:
The SMSM callback processing work is scheduled in a single thread workqueue
rather than global workqueue. The SMSM callback processing work holds the mutex
until it extracts the Wakelock usage information from the kfifo.

Change-Id: I516db9c5fed82482494ffb25d62b8ea35abb35ab
Reviewed-on: http://gerrit.pcs.mot.com/399236
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Yuanyuan Zhong <a6510c@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>