ARM: 7400/1: vfp: clear fpscr length and stride bits on entry to sig handler
ARM: 7400/1: vfp: clear fpscr length and stride bits on entry to sig handler

The ARM PCS mandates that the length and stride bits of the fpscr are
cleared on entry to and return from a public interface. Although signal
handlers run asynchronously with respect to the interrupted function,
the handler itself expects to run as though it has been called like a
normal function.

This patch updates the state mirroring the VFP hardware before entry to
a signal handler so that it adheres to the PCS. Furthermore, we disable
VFP to ensure that we trap on any floating point operation performed by
the signal handler and synchronise the hardware appropriately. A check
is inserted after the signal handler to avoid redundant flushing if VFP
was not used.

Reported-by: Peter Maydell <peter.maydell@linaro.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>
Change-Id: I38e109bcff031f11103469bb9fff8d260911cdf1
Reviewed-on: http://gerrit.pcs.mot.com/437166
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>