IKHSS7-26799: Read UTAG and Lock Battery Meter
This change is needed for Customer RF Testing.
It is needed to prevent the phone from shutting down when RF Testing is
conducted at 3.4 Volts.  With this change if UTAG battery=meter_locked
then the Battery meter will show a ? and always stay at 50%.  It is expected
that only RF testing units will set this UTAG.

Change-Id: Ie578b21d25d4f39680128beb8a81b0c950b3f09f
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/400016
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>