base: genlock: allow synchronization with a single gralloc handle
In order to support synchronization in a process with a single
gralloc handle we require the ability to write lock a buffer
while it is already read locked by the same handle. This change
extends the concept of an exclusive write lock or recursive read
locks to a genlock handle (rather than the genlock lock).
Genlock cannot provide deadlock protection because the same
handle can be used simultaneously by a producer and consumer.
In practice an error will still be generated when the timeout
expires.

CRs-fixed: 356263
Change-Id: Ic5a3ca3ce83eca0b86b92459230c259a99f352a8
Signed-off-by: Jeff Boody <jboody@codeaurora.org>
Signed-off-by: Zhoulu Luo <zluo@codeaurora.org>
Reviewed-on: http://gerrit.pcs.mot.com/416641
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Sang Hyon Ji <w21725@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>