IKHSS7-27400 : Avoid NULL pointer panic in rndis_response_complete
There was a panic seen in which the rndis pointer was NULL in the
ECONNRESET/ESHUTDOWN case. Adding a check to ensure the panic
doesn't happen again. Also a warning was added to get more debug
info as we aren't expecting the rndis pointer to be NULL in this
function. rndis is reassigned when tethering is re-enabled so this
change is safe.

Change-Id: I072323b83d0eee4b722810a9047d7001f6966f82
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/405998
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Ravikumar Vembu <a18394@motorola.com>