IKHSS7-29586: Charge State Stuck in Charge Fail
Randomly the Charge Fail is occuring without an interrupt or a missed
interrupt occurs and the Charge FAil Clear is not set.  This means that
the Phone will not charge unless the battery is pulled or Charge Fail
bit is manually set.

This change sets the Charge Fail Bit during the Charging Drivers Probe and
also when an Accessory is inserted or removed during the Normal Temperature
State(0C-45C).

Change-Id: I5b8bc1d7b8a92f5bbc4f3f16cce7d25613fab630
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/410299
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>