IKHSS7-31077 : Re-Authenticate the dock when PPD -> SPD OR SPD -> PPD  transitions happen
    The issue is seen when  CHARGER  which is inserted into the dock
is unplugged at MAINS,and plugged in immediately. In the past we had
issues where we would see multiple IDGND interrupts, when a audio cable
is inserted slowly into the CAR dock.  The solution was to debounce
sporadic ID_GND interrupts as seen with a car dock.

    This is in turn breaking the use case mentioned above where we
get 2 IDGND interrupts. The second ID_GND we receive as a result of
applying the CHARGER is treated as a sporadic ID_GND interrupt.

    The key in this change is the delay of 3s. When the VBUS drops,
and if DP is pulled up immediately, a SE1 condition is detected.
As a result, the VBUS should be allowed to settle down, before the
DP is pulled up.

Change-Id: Ib3b6099c58a1465fb89f970b627bba3014afcac0
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>

AAA

Change-Id: I21c09e2ddbf9e4e46c5d735dae5f840f4e96bd87
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Change-Id: Ib3b6099c58a1465fb89f970b627bba3014afcac0
Reviewed-on: http://gerrit.pcs.mot.com/411164
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>