IKHSS7-29704: Force L14 ON when Accy is attached.
The Coulomb Counter supply is getting turned on and off when on a charger.
When the Battery gets full this leads to an offset that gets applied to the
Accumulator and presents it self as a 20 mA drain.

This change requests L14, the Coulomb Counter Supply, to the ON state whenever
an accy is attached and OFF when no acy is attached.

Change-Id: I191eee7bd5512b25c3952d31e73cec9bf6c9b80a
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/415685
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Vidyadhar Gurram <Vidyadhar.Gurram@motorola.com>
Reviewed-by: Timothy McCune <w12222@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>