IKHSS7-36579:Adjust Remaining Charge from Voltage and Use Aged Capacity
Some Phones are shutting down around 13%.  This fix attempts at preventing it.

The change gradually speeds up the meter if it detects a the Battery Voltage,
which has been current compensated, crosses the 5% OCV mark. Also to prevent
the scenario we will be feeding the Aged Capacity Calculated Value back into
the scale factor so that once the issue is seen it will not happen on the next
iterations.

Change-Id: I0d2d5f5a88dc08abbbbaa1043064a57b2f23cce0
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/430065
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Vidyadhar Gurram <Vidyadhar.Gurram@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>