IKHSS7-37622: Handle PPD to SPD transition in the process of Authentication.
    If the CHARGER is inserted into a UART Authentication dock(DESK/CAR) while the
Authentication is in progress, the phone assumes that it was a erroneous ID interrupt.
As a result, both the phone & CHARGER would be sourcing the power to the DOCK.

    In order to avoid such condition to happen, we check whether a CHARGER is connected
to the dock when the authentication PASSES and if YES, we disable the RVRS mode.

    As a part of this change, the RVRS mode is also turned OFF if the authentication
fails because we only support Audio if authentication passes.

Change-Id: I6e8dd05ad7e9995e2fcf330e5dfb34513e404f72
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/433433
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>