IKHSS7-38563: Switch non-EMU path to USB mode instead of UART.
    Currently the Accessory detection driver switches the EMU path
and non-EMU path between audio & UART mode respectively. As a result
there would be a bias voltage on the D+ line because of the PULL UP
resistor that we have in our configuration.

    One solution would have been to have no PULL UP resistor on D+ line
when the non-EMU path was choosen by Audio driver. This needs a change in
device tree across all the products and all HW versions.

    Instead, this solution would choose non-EMU path as USB instead of
UART because the voltage would be 0V when it is in unenumerated state.

Change-Id: Iff59a9b3e81c3bb14699bcbdd15a754d141a4613
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/436970
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>